chatbot_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
streamlit==1.32.2
